+++ src/controller/RecipeController.js
@@ -1,0 +1,69 @@
+const recipesRepository = require('../Repository/RecipesRepository');
+const userRepository = require('../Repository/UserRepository');
+const Ingredient = require('../models/Ingredient');
+
+function addRecipe(req, res, next){
+    let ingredients = [];
+    req.body.ingredients.forEach((ingredient) => {
+        let i = new Ingredient(ingredient.name, ingredient.unit, ingredient.quantity);
+        if(!i.verifyAttributes())
+            return res.json({error: true});
+
+        ingredients.push(i);
+    });
+
+    if(ingredients.length) {
+        userRepository.getUser(req.body.auth).then((response) => {
+            if(response.user.email) {
+                recipesRepository.addRecipe(response.user.email, req.body.name, req.body.time, req.body.portions, req.body.visibility, req.body.steps).then((response) => {
+                    if(response.error)
+                        return res.json(response);
+
+                    recipesRepository.addIngredients(req.body.ingredients, response.recipeid).then((response) => {
+                        if(response.error)
+                            return res.json(response);
+
+                        else if(req.body.categories && req.body.categories.length) {
+                            recipesRepository.addCategories(req.body.categories, response.recipeid).then(response => {
+                                res.json(response);
+                            });
+                        }
+
+                        else
+                            res.json({error: false});
+                    });
+                });
+            }
+
+            else
+                res.json({error: true});
+        });
+    }
+}
+
+function deleteRecipe(req, res, next){
+    var body = req.body;
+    
+    userRepository.getUser(body.auth).then((response) => {
+        if(response.user)
+            recipesRepository.deleteRecipe(body.recipeid, response.user).then(response => res.json(response));
+
+        else
+            res.json({error: true, details: 'An error occurred while fetching the user. User not found.'});
+    });
+}
+
+function getRecipeById(req, res, next){
+    var body = req.body;
+    
+    recipesRepository.getRecipe(body.recipeId, body.auth).then(response => res.json(response));
+}
+
+function getRecipeList(req, res, next){
+    var body = req.body;
+    
+    recipesRepository.getAllRecipes(body.auth).then(response => res.json(response));
+
+}
+
+module.exports = { addRecipe, deleteRecipe, getRecipeById, getRecipeList };
